backend/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database dependencies
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# AI/ML dependencies
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
scikit-learn==1.3.2
ultralytics==8.0.196
mediapipe==0.10.7

# Image processing and computer vision
opencv-contrib-python==4.8.1.78
imutils==0.5.4
scipy==1.11.4

# Data handling
pandas==2.1.3
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP and API clients
httpx==0.25.2
requests==2.31.0
aiofiles==23.2.1

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
typer==0.9.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
